+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix extern test declaration. This should fix issue #2
+- Fix minor GCC warnings
+
 ## [1.4.0] - 2020-03-12
 
 ### Added